Completed 2022-12-11
Probably could check for cycles in part b, but since all the monkeys
have prime-numbered conditions, I'd guess that those would be longer
than the 10000 rounds, unless something happens to work out nicely
from the operations of each monkey.  As it is, it runs decently fast
anyways, so no need to optimize.

|             | Debug  | Release |
|--------------------------------|
| Example     | 96 ms  | 10 ms   |
| PuzzleInput | 366 ms | 34 ms   |